update: strtategy

+++ src/models/bar.rs
@@ -3,21 +3,21 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Bar {
     #[serde(rename = "c")]
-    pub close_price: f32,
+    pub close_price: f64,
     #[serde(rename = "h")]
-    pub high_price: f32,
+    pub high_price: f64,
     #[serde(rename = "l")]
-    pub low_price: f32,
+    pub low_price: f64,
     #[serde(rename = "n")]
     pub n: u32,
     #[serde(rename = "o")]
-    pub open_price: f32,
+    pub open_price: f64,
     #[serde(rename = "t")]
     pub timestamp: String,
     #[serde(rename = "v")]
-    pub volume: u32,
+    pub volume: u64,
     #[serde(rename = "wv")]
-    pub wv: f32,
+    pub wv: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
